requirements.txt
bunch==1.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
facebookads==2.11.4
Flask==1.0.2
Flask-Excel==0.0.7
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsons==0.5.3
lml==0.0.4
MarkupSafe==1.0
numpy==1.15.2
pandas==0.23.4
psycopg2==2.7.5
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
texttable==1.4.0
urllib3==1.23
Werkzeug==0.14.1
xlrd==1.1.0
